+++ demo/assets/js/app.js
@@ -20,6 +20,7 @@ import "phoenix_html";
 // Establish Phoenix Socket and LiveView configuration.
 import { Socket } from "phoenix";
 import { LiveSocket } from "phoenix_live_view";
+import { hooks as colocatedHooks } from "phoenix-colocated/doggo";
 import topbar from "../vendor/topbar";
 
 import "../css/app.scss";
@@ -29,6 +30,7 @@ let csrfToken = document
   .getAttribute("content");
 let liveSocket = new LiveSocket("/live", Socket, {
   params: { _csrf_token: csrfToken },
+  hooks: colocatedHooks,
 });
 
 // Show progress bar on live navigation and form submits