+++ lib/doggo/storybook/shared.ex
@@ -130,6 +130,45 @@ defmodule Doggo.Storybook.Shared do
     """
   end
 
+  defp icon_svg(:pause) do
+    """
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      class="lucide lucide-pause"
+    >
+      <rect x="14" y="4" width="4" height="16" rx="1" />
+      <rect x="6" y="4" width="4" height="16" rx="1" />
+    </svg>
+    """
+  end
+
+  defp icon_svg(:play) do
+    """
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      class="lucide lucide-play"
+    >
+      <polygon points="6 3 20 12 6 21 6 3" />
+    </svg>
+    """
+  end
+
   defp icon_svg(:chevron_right) do
     """
     <svg